Use cast to int to avoid unnamed types

+++ src/parser_urdf.cc
@@ -1755,7 +1755,7 @@ void CreateGeometry(TiXmlElement* _elem,
       }
       break;
     default:
-      sdfwarn << "Unknown body type: [" << _geom->type
+      sdfwarn << "Unknown body type: [" << static_cast<int>(_geom->type)
         << "] skipped in geometry\n";
       break;
   }
@@ -2437,7 +2437,7 @@ void CreateJoint(TiXmlElement *_root,
         jtype = "fixed";
         break;
       default:
-        sdfwarn << "Unknown joint type: [" << _link->parent_joint->type
+        sdfwarn << "Unknown joint type: [" << static_cast<int>(_link->parent_joint->type)
           << "] in link [" << _link->name << "]\n";
         break;
     }
